nobody/requirements.txt
# web app
Django==1.6.6
# ipython==2.1.0
MySQL-python==1.2.5
django-mptt==0.6.1
django-annoying==0.8.0
django-floppyforms==1.2.0
ajaxuploader==0.3.8
redis==2.10.3
hiredis==0.1.5
lxml==3.1.2
django-overextends==0.3.2
djangorestframework==2.4.2
thrift==0.9.1
Fabric==1.10.0
django-rest-swagger==0.1.14
raven==5.1.1
xlrd==0.9.3
django-redis-cache==0.13.0
Pillow==2.6.1
# numpy==1.9.0
# scipy==0.14.0
pandas==0.15.0
ipdb==0.8
requests==2.4.3
celery[redis]==3.1.17
python-logstash==0.4.2
# protobuf==2.6.1
addict==0.2.7
shortuuid==0.4.2
wechat-sdk==0.5.5
scrapy==0.24.6
service_identity==14.0.0
rsa==3.1.4
xlwt==1.0.0
bleach==1.4.1
python-slugify==1.1.2
gunicorn==19.3.0
setproctitle==1.1.8
parsimonious==0.6.2
mammoth==0.3.20
openpyxl==2.2.5
# analysis
appnope==0.1.0
backports.ssl-match-hostname==3.4.0.2
certifi==2015.4.28
decorator==4.0.2
funcsigs==0.4
functools32==3.2.3.post2
gnureadline==6.3.3
ipykernel==4.0.3
ipython==4.0.0
ipython-genutils==0.1.0
ipywidgets==4.0.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.0.0
jupyter-console==4.0.1
jupyter-core==4.0.4
MarkupSafe==0.23
matplotlib==1.4.3
mistune==0.7.1
mock==1.3.0
nbconvert==4.0.0
nbformat==4.0.0
nose==1.3.7
notebook==4.0.4
numpy==1.9.2
path.py==8.1
pbr==1.6.0
pexpect==3.3
pickleshare==0.5
ptyprocess==0.5
PyBrain==0.3
Pygments==2.0.2
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
pyzmq==14.7.0
qtconsole==4.0.1
scikit-learn==0.16.1
scipy==0.16.0
simplegeneric==0.8.1
six==1.9.0
sklearn==0.0
terminado==0.5
tornado==4.2.1
traitlets==4.0.0
wheel==0.24.0
